Move stores to IO namespace.
