+++ osu.Framework/Audio/AudioManager.cs
@@ -7,7 +7,6 @@ using ManagedBass;
 using osu.Framework.Audio.Sample;
 using osu.Framework.Audio.Track;
 using osu.Framework.Configuration;
-using osu.Framework.Resources;
 using osu.Framework.Threading;
 
 namespace osu.Framework.Audio
