+++ osu.Framework/Audio/Sample/SampleManager.cs
@@ -1,7 +1,7 @@
 ﻿//Copyright (c) 2007-2016 ppy Pty Ltd <contact@ppy.sh>.
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-using osu.Framework.Resources;
+using osu.Framework.IO.Stores;
 
 namespace osu.Framework.Audio.Sample
 {
