+++ osu.Framework/Graphics/Shaders/ShaderManager.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using osu.Framework.Logging;
-using osu.Framework.Resources;
 
 namespace osu.Framework.Graphics.Shaders
 {
