+++ osu.Framework/Graphics/Shaders/ShaderPart.cs
@@ -11,7 +11,6 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using osu.Framework.Resources;
 
 namespace osu.Framework.Graphics.Shaders
 {
