+++ osu.Framework/Graphics/Textures/PrefixTextureStore.cs
@@ -2,7 +2,7 @@
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
 using System.Resources;
-using osu.Framework.Resources;
+using osu.Framework.IO.Stores;
 
 namespace osu.Framework.Graphics.Textures
 {
