+++ osu.Framework/IO/Stores/ChangeableResourceStore.cs
@@ -2,9 +2,8 @@
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
 using System;
-using System.Resources;
 
-namespace osu.Framework.Resources
+namespace osu.Framework.IO.Stores
 {
     public class ChangeableResourceStore<T> : ResourceStore<T>
     {
