+++ osu.Framework/IO/Stores/NamespacedResourceStore.cs
@@ -1,13 +1,7 @@
 ﻿//Copyright (c) 2007-2016 ppy Pty Ltd <contact@ppy.sh>.
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace osu.Framework.Resources
+namespace osu.Framework.IO.Stores
 {
     public class NamespacedResourceStore<T> : ResourceStore<T>
     {
