+++ osu.Framework/osu.Framework.csproj
@@ -186,14 +186,14 @@
     <Compile Include="Graphics\Containers\ProcessingContainer.cs" />
     <Compile Include="Graphics\Drawables\Box.cs" />
     <Compile Include="Graphics\Drawable_Interaction.cs" />
-    <Compile Include="Resources\CachedResourceStore.cs" />
-    <Compile Include="Resources\ChangeableResourceStore.cs" />
-    <Compile Include="Resources\DllResourceStore.cs" />
-    <Compile Include="Resources\FileSystemResourceStore.cs" />
-    <Compile Include="Resources\GlyphStore.cs" />
-    <Compile Include="Resources\IResourceStore.cs" />
-    <Compile Include="Resources\NamespacedResourceStore.cs" />
-    <Compile Include="Resources\ResourceStore.cs" />
+    <Compile Include="IO\Stores\CachedResourceStore.cs" />
+    <Compile Include="IO\Stores\ChangeableResourceStore.cs" />
+    <Compile Include="IO\Stores\DllResourceStore.cs" />
+    <Compile Include="IO\Stores\FileSystemResourceStore.cs" />
+    <Compile Include="IO\Stores\GlyphStore.cs" />
+    <Compile Include="IO\Stores\IResourceStore.cs" />
+    <Compile Include="IO\Stores\NamespacedResourceStore.cs" />
+    <Compile Include="IO\Stores\ResourceStore.cs" />
     <Compile Include="Graphics\Textures\Texture.cs" />
     <Compile Include="Graphics\Textures\TextureAtlas.cs" />
     <Compile Include="Graphics\Textures\TextureLoader.cs" />
